create page-specific nav components
