+++ apps/website/src/components/clips/ClipsNavigation.tsx
@@ -1,0 +1,17 @@
+import { PageNavigation } from "../shared/PageNavigation";
+
+export default function ClipsNavigation() {
+  const clipsNavigationItems = [
+    {
+      href: "/clips/",
+      label: "Clips",
+      exact: true,
+    },
+    {
+      href: "/clips/submit-clip",
+      label: "Submit Clip",
+    },
+  ];
+
+  return <PageNavigation navItems={clipsNavigationItems} />;
+}
